+++ eth-test-helpers/src/setup.ts
@@ -1,3 +1,8 @@
+/**
+ * This file provides utility functions related to test setup, such as creating a test provider,
+ * optimizing test times via snapshots, and making test accounts.
+ */
+
 import {
   RevertTraceSubprovider,
   SolCompilerArtifactAdapter,
@@ -10,12 +15,14 @@ import {
 import { ethers } from 'ethers'
 import * as path from 'path'
 import { makeDebug } from './debug'
+import { createFundedWallet } from './wallet'
+
 const debug = makeDebug('helpers')
 
 /**
  * Create a test provider which uses an in-memory, in-process chain
  */
-export function makeTestProvider(): ethers.providers.JsonRpcProvider {
+export function provider(): ethers.providers.JsonRpcProvider {
   const providerEngine = new Web3ProviderEngine()
   providerEngine.addProvider(new FakeGasEstimateSubprovider(5 * 10 ** 6)) // Ganache does a poor job of estimating gas, so just crank it up for testing.
 
@@ -50,7 +57,7 @@ export function makeTestProvider(): ethers.providers.JsonRpcProvider {
  * @param provider The provider that's used within the tests
  * @param cb The callback to execute that generates the state we want to snapshot
  */
-export function useSnapshot(
+export function snapshot(
   provider: ethers.providers.JsonRpcProvider,
   cb: () => Promise<void>,
 ) {
@@ -86,4 +93,64 @@ export function useSnapshot(
       d('recreated snapshot id:%s', snapshotId)
     }
   }
+}
+
+export interface Roles {
+  defaultAccount: ethers.Wallet
+  oracleNode: ethers.Wallet
+  oracleNode1: ethers.Wallet
+  oracleNode2: ethers.Wallet
+  oracleNode3: ethers.Wallet
+  oracleNode4: ethers.Wallet
+  stranger: ethers.Wallet
+  consumer: ethers.Wallet
+}
+
+export interface Personas {
+  Default: ethers.Wallet
+  Neil: ethers.Wallet
+  Ned: ethers.Wallet
+  Nelly: ethers.Wallet
+  Carol: ethers.Wallet
+  Eddy: ethers.Wallet
+}
+
+interface Users {
+  roles: Roles
+  personas: Personas
+}
+
+/**
+ * Generate roles and personas for tests along with their corrolated account addresses
+ */
+export async function users(
+  provider: ethers.providers.JsonRpcProvider,
+): Promise<Users> {
+  const accounts = await Promise.all(
+    Array(8)
+      .fill(null)
+      .map(async (_, i) => createFundedWallet(provider, i).then(w => w.wallet)),
+  )
+
+  const personas: Personas = {
+    Default: accounts[0],
+    Neil: accounts[1],
+    Ned: accounts[2],
+    Nelly: accounts[3],
+    Carol: accounts[4],
+    Eddy: accounts[5],
+  }
+
+  const roles: Roles = {
+    defaultAccount: accounts[0],
+    oracleNode: accounts[1],
+    oracleNode1: accounts[2],
+    oracleNode2: accounts[3],
+    oracleNode3: accounts[4],
+    oracleNode4: accounts[5],
+    stranger: accounts[6],
+    consumer: accounts[7],
+  }
+
+  return { personas, roles }
 }
